[#27991] YSQL: Reduce number of memory allocations in case of using PgDmlRead::SetHoldingRequestedYbctids method
Summary:
The `PgDmlRead::SetHoldingRequestedYbctids` methods accepts vector which is not reused. It is reasonable to use generator instead.

Also there are several unnecessary copies of the `HoldingYbctidProvider` which is removed in the diff.
The `HoldingYbctidProvider` it self is also changed a bit, since it is no more that `SimpleYbctidProvider` but with vector owning.
Jira: DB-17612

Test Plan: Jenkins

Reviewers: aagrawal, amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45417