[PLAT-18117] PIT restore from a backup should compare with earliest snapshot time
Summary:
When trying to find which backup to use for restoring to the given PIT, we were originally using
the backup create time. While this will work for cases where the retention window is large, it fails
for cases with small retention window as the snapshot create time may itself be more that window.

To fix this, instead of using backup create time we now use a new field `firstSnapshotTime` to find the
earliest usable backup.

Changes in this PR also include addition of a new column to the backup table `first_snapshot_time` and
another migration to populate it for cases where backup retention window exists.

Test Plan:
Added UTs which fetch the usable backup in multiple scenarios
Tested migration with portal backups

Reviewers: agunta, dshubin

Reviewed By: dshubin

Differential Revision: https://phorge.dev.yugabyte.com/D45444