[PLAT-18074] Add subtask execution time such that it can displayed on the UI
Summary:
This records the queuing and execution times taken by subtasks. The existing API endpoint is refactored to increase performance by not getting for all targets and tasks. It also fixes the retryability flag.
For old tasks before this change, the times will be -1 (default). So, any time < 0 can be ignored as not available.

Test Plan: Manually tested.

Reviewers: amalyshev, rmadhavan, kkannan

Reviewed By: amalyshev, rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D45452