[#26425] CDC: Integrate polling sys catalog tablet in virtual WAL
Summary:
In this revision, changes are introduced to poll the sys catalog tablet in the virtual WAL so as to identify change in the publication's tables list. To accurately determine any change to the publication, be it in the form of `ALTER PUBLICATION` or due to the creation of new tables (in ALL TABLES publications) we need CHANGE records from two catalog tables, namely `pg_publication_rel` and `pg_class`. At the time of stream creation, these tables are added to the appropriate stream metadata data structures.

During InitVirtualWAL, walsender will send the list of publications and a boolean which tells us if any of those publications is an ALL TABLES publication. A tablet queue will be created for the sys catalog tablet. This tablet queue will be treated like any other tablet queue and GetChanges will be called on sys catalog tablet whenever this queue becomes empty. Any record popped from this queue will be considered a publication refresh record (having `VWALRecordType::PUBLICATION_REFRESH`). This will ensure that whenever commit time ties are encountered, the records from sys catalog tablet will come last in the sorted order. When a record belonging from sys catalog tablet queue will be popped from the priority queue, we will check the following:

  - If the record is from pg_class table, then check whether this VWAL polls on an ALL TABLES publication and the said record is an INSERT with relkind relation (i.e. this record indicates that a new table has been created).
  - If the record is from pg_publication_rel, then check whether the record is a DELETE  record or if its an INSERT indicating addition of table in one of publications this VWAL is polling.
If any one of the above two conditions is true, then we signal walsender to invalidate rel cache and get the table list from publication as of the commit_time of the record popped from the catalog tablet queue. This list will then be relayed back to Virtual WAL which will make appropriate adjustments in its data structures.

##### Upgrade / rollback safety
All the changes for this feature will be guarded by a default false test flag `TEST_ysql_yb_enable_implicit_dynamic_tables_logical_replication` until the development is complete. This flag will then be made a preview flag and after sufficient bake time be promoted to a default true auto flag.

  - Two new fields, a list of publication oids  (`publication_oid`) and a boolean `pub_all_tables` have been added to the `InitVirtualWALForCDCRequestPB`

This diff does not handle the upgrade of existing streams to a version where the flag `ysql_yb_enable_implicit_dynamic_tables_logical_replication` is turned on by default. The upgrade path will be provided in a subsequent revision. This feature is not rollback safe.
Jira: DB-15788

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testDynamicTableAdditionForTablesCreatedBeforeStreamCreationWithPubRefresh'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testDynamicTableAdditionForTablesCreatedBeforeStreamCreationWithoutPubRefresh'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testDynamicTableAdditionForAllTablesPublicationWithPubRefresh'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testDynamicTableAdditionForAllTablesPublicationWithoutPubRefresh'
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestStreamIndependenceWhilePollingCatalogTablet

Reviewers: skumar, asrinivasan, stiwary

Reviewed By: asrinivasan

Subscribers: yql, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45472