[PLAT-18078][dr][ddl] Show proper database name for sequences_data table
Summary:
This diff sets the database name based on the table id for the sequences_data table.

There is no postgres schema for that table and it belongs to the whole database and thus we put '_' as the pg schema.

Test Plan: Created an automatic mode DR config with multiple databases in replication and made sure the proper database name is shown for each sequences_data replication stream.

Reviewers: amindrov, spothuraju, jmak, nbhatia, yng

Reviewed By: spothuraju

Differential Revision: https://phorge.dev.yugabyte.com/D45325