[BACKPORT 2025.1][PLAT-18120] Bump up YBC client and server version to 2.2.0.2-b9
Summary:
Original PR: 58ad8cc0d49664889a5a6a179ec7ddd3f628f775 / D45466
Diff updates the YBC version on YBA to 2.2.0.2-b9. This adds suport for backups
during DDL. It is currently behind a feature flag: yb.backup.enable_backups_during_ddl
which is turned off by default.

YBC commit: https://github.com/yugabyte/ybc/commit/0a5552271cb5359d282241952f6d3441c0e3a0fd

```
On top of read-time support, there are a few contract changes for snapshot
creation and export which needs to be handled on YBC.

- For DDL support backups, snapshot does not include table names for YSQL. We
only need to provide a single entry with the namespace name and ID. Tables will be
figured on the master side.
- Add flag `include_ddl_in_progress_tables` to export snapshot call.

Both of these changes are introduced in the flow based on `read-time-supported`.

* [PLAT-18041]YBC-YBDB contract changes to support backups during DDL

On top of read-time support, there are a few contract changes for snapshot
creation and export which needs to be handled on YBC.

- For DDL support backups, snapshot does not include table names for YSQL. We
only need to provide a single entry with the namespace name and ID. Tables will be
figured on the master side.
- Add flag `include_ddl_in_progress_tables` to export snapshot call.

Both of these changes are introduced in the flow based on `read-time-supported`.
```

Test Plan: dev itest

Reviewers: dshubin

Reviewed By: dshubin

Differential Revision: https://phorge.dev.yugabyte.com/D45495