[#28034] CDC: Handle getting consistent stream safe time for non transactional YCQL tables
Summary:
In YCQL, by default, distributed transactions are disabled on the tables. The tablets of such tables do not have a transaction participant. Because of this when CDC tries to fetch the min start time among all running transactions it receives a `kInvalid` value. CDC producer wrongly concludes that loading of transactions is in progress and does not send any record in that GetChanges call. In this way when CDC is run for non-transactional YCQL tables, no records are shipped from the CDC producer as it keeps waiting for loading of transactions to complete.

In this revision we fix this issue by returning `kMax` as the min start time among all running transactions for non-transactional tables. CDC producer interprets `kMax` as no running transactions on the tablet. This is correct since such tables only have single shard transactions and hence can never have any running transactions.

Jira: DB-17655

Test Plan:
Ran all the connector tests in `YugabyteDBCQLTest` against a yugabyted instance with these changes.
Existing CDC ctests.

Reviewers: skumar, asrinivasan

Reviewed By: skumar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45505