[BACKPORT 2025.1.0][#28053] docdb: Increase default ysql lease ttl
Summary:
Bumps the default YSQL lease TTL to 5 minutes.

We decided to release the YSQL lease ahead of the table locking feature to kick the tires on it. Table locking will not be released on the 2025.1 series, but the YSQL lease is enabled on this series. The significance of the lease TTL without table locking is tservers will kill all hosted pg sessions when they are unable to communicate with the master leader for the lease TTL. We want to minimize negative impact on clusters due to the YSQL lease so we're bumping up the default TTL on the 2025.1 branch.
Jira: DB-17678

Original commit: 3387f8d84230d659424c01ef0d5f83d78959511d / D45548

Test Plan: n/a

Reviewers: amitanand, bkolagani

Reviewed By: bkolagani

Subscribers: ybase, slingam, rthallam

Differential Revision: https://phorge.dev.yugabyte.com/D45557