[PLAT-18165][PLAT-18175] UpgradeYBC task did not update universe details.
Summary:
If YBC gets upgraded via the UpgradeYbc subtask, and not the background thread,
the universeDetails never get updated. This was still handled as the background thread
would eventually detect it, and update the universe details.
But, on YBM, we have the background thread disabled. This would then cause every backup to first update YBC.
On large universes, this was adding significant and noticeable time to customer backups

Also updated the subtask to correctly handle when a node is offline, as it shouldn't update the universeDetails.ybcVersion
util all nodes get upgraded.

Finally, enabled the backgroudn task in YBM, with a seperate, non-ybm flag to turn it off

Test Plan:
updated unit tests to confirm update
Deployed YBA in 'cloud mode' and validated ybc gets upgraded and universe details updated

Tested new flag enables/disables background task

Validated the ybc upgrade during create backup with:
1. all nodes online, universe is updated
2. 1 node offline. ybc updated on 2 nodes, universe details not updated. backup succeeds (slowly)
3. all nodes back online. ybc updated and universe details updated

Reviewers: vkumar, vbansal

Reviewed By: vbansal

Subscribers: vbansal

Differential Revision: https://phorge.dev.yugabyte.com/D45563