[#28059] DocDB: Set leader_term when processing deleted tablet in transaction coordinator
Summary:
It could happen that involved tablet was deleted when transaction is being applied.
In this case we remove this tablet from pending.
Sometimes it is the last pending tablet, so we should replicate the fact that transaction was applied in all tablets.
But postponed leader actions were not initialized in related codepath.

This diff addresses the issue by utilising CoordinatorLock for this code path.

This bug is not too critical since this sitation is pretty rare and we still will replicate the fact that transaction was applied in all involved tablets via Poll mechanism.
Jira: DB-17688

Test Plan: Jenkins

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45554