[#23721] docdb: Remove table info objects from table map upon deletion.
Summary:
When `CleanUpDeletedTables` was originally introduced in 80de98cc662ddb55aa50ba8e100044284613f0b7, it removed the table ids of the deleted table from the map of table ids. At some point along the way this behaviour was removed. This diff restores that behaviour. The motivation is to allow a master process to garbage collect deleted tables, so users which continually create and delete tables won't OOM the master leader process. In order to enable this we also need to delete the tablet info objects, as the ownership model is `TabletInfo` holds a `scoped_refptr<TableInfo>`, keeping the tablet's `TableInfo` object alive until the `TabletInfo` itself is freed.

History:
11687382fe22c740a731ea774570111f075fe393 stopped removing deleted tables from the in memory maps. A [[ https://github.com/yugabyte/yugabyte-db/issues/2721 | bug ]] was causing crashes because roll forward for deletion wasn't implemented. That was implemented in d45058e0369310fa0920070d98d9da35e30d7bc9. However the logic to remove deleted tables from the map was never added back, I wasn't able to determine why.

One would hope there'd be some discussion in D9227, but no.

**Upgrade/Rollback safety:**
This diff adds a new enum type to the status object. It is only used internally to differentiate a deleted tablet from a missing tablet. It is possible for this value to be returned to remote clients.
Jira: DB-12629

Test Plan:
```
% ./yb_build.sh fastdebug --cxx-test delete_table-test --gtest_filter 'RemoveTablesFromMapsTest.RemoveHiddenTableFromMap:RemoveTablesFromMapsTest.RemoveDeletedTableFromMap'
```

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: hsunder, slingam, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D45214