[BACKPORT 2024.2][PLAT-18174] Break out of infinite loop for Ybc upgrades by removing deleted universes
Summary:
Original commit: 75df7707e39be662d162443bcd43a7c09acc26d9 / D45595
On dev-portal we noticed high cpu usage by YbcUpgrade thread. This was happening because
universes to upgraded are added to `ybcUpgradeUniverseSet` and never removed if they are deleted.
Also this was a `while` loop as opposed to being a simple if check.

## Why don't we need the `while` loop anymore ?

This is because we are already doing a `waitForYbc` call which has 20 retries. So we don't
explicitly need to retry the `pollUpgradeTaskResult` again. This is the same as the K8s universes case.

## Fix

- Modifying the while loop to an if conditional
- Removing the universe from `ybcUpgradeUniverseSet` once its deleted.

Test Plan: No tests

Reviewers: nsingh, vbansal

Reviewed By: nsingh

Differential Revision: https://phorge.dev.yugabyte.com/D45625