[#27061] YSQL: Use one transaction unit for all COPY FROM commands in a transaction block
Summary:
Before this fix, a transaction block with COPY commands could be split
into multiple transactions as follows:

```
BEGIN;
COPY FROM ... ROWS_PER_TRANSACTION 10; ----> assume this is a COPY of 25 rows. There would be 3 transactions of 10, 10 and 5 rows each one after the other. Each is committed after the rows are inserted.
COPY FROM; ----> a new transaction is started at this point. All remaining COPY commands are part of this second transaction.
COPY FROM;
...
COMMIT;
```

We intend to ignore the `ROWS_PER_TRANSACTION` setting when
COPY is part of a transaction block. However, instead of checking
`IsTransactionBlock()`, it seems like `YbIsDataSent()` was used as a
proxy to figure if we are in a transaction block (refer fe824215). Note that
`YbIsDataSent()` is false for the first statement in a transaction block (we
ignore the trivial data sent on a `BEGIN;` statement and avoid marking
`ybDataSent`).

The fix is to ignore `ROWS_PER_TRANSACTION` by checking if
`IsTransactionBlock()` is true, instead of `YbIsDataSent()`.

### NOTE:

(1) The same issue exists for a batch of COPY commands too. `ROWS_PER_TRANSACTION`
is honored for the first statement, and ignored for the remaining COPY commands because
`YbIsDataSent()` becomes true after the first COPY command.

(2) If the first statement in a transaction block was something other than COPY,
then everything would be part of a single transaction unit even without this fix.

(3) If `ROWS_PER_TRANSACTION` was larger than the number of rows in the first COPY,
the COPY would still commit when the statement ends. This is the behaviour when
`ROWS_PER_TRANSACTION` is honored.
Jira: DB-16542

Test Plan:
./yb_build.sh release --cxx-test pgwrapper_pg_copy-test --gtest_filter PgCopyTest.TestCopyAtomicTxnBlock
./yb_build.sh release --java-test org.yb.pgsql.TestPgBatch#testCopyFromInBatch

Reviewers: smishra, patnaik.balivada, huapeng.yuan

Reviewed By: patnaik.balivada

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D45646