[#28144] YSQL: Let auth backend use tserver cache with connection manager
Summary:
When connection manager is used, there can be many auth backends created.
Currently in order to perform authorization check, an auth backend still needs
to send one RPC to master to load a number of shared relations (e.g., pg_authid,
pg_auth_members, pg_database). There can be very large number of logical
connection churns which means there will be many auth backends come and go. If
we can use tserver cache for auth backends during its authorization check, we
can significantly reduce the master load.

This diff makes auth backend to use tserver cache to do its work. The high level idea
is that the shared memory catalog version (which is an approximation of the latest
master catalog version) is used to perform the prefetching work. A new function
`YbUseTserverResponseCacheForAuth` is added to check whether tserver cache can
be used during authorization check. A second change is required to increment
catalog version if a global impact DDL is detected. In this case a DDL writes to
shared relation but it may not be classified to increment catalog version. For
example, when a user's password is changed, currently we do not increment
catalog version. But if the stale password is stored in tserver cache, then the
new password will not be loaded if the catalog version isn't incremented.

I have found that each auth backend currently performs 3 RPCs. After this diff,
the number of master RPCs is reduced to 2.

New unit tests are added to verify the reduction of master RPCs and global
impact DDL statements always increment catalog version to ensure stale cache
entries in tserver cache are no longer used.
Jira: DB-17786

Test Plan:
./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionConnManagerTest.TestConnectionManagerRpcCount/0
./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionConnManagerTest.TestConnectionManagerRpcCount/1
./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionConnManagerTest.TestConnectionManagerChangePassword/0
./yb_build.sh release --cxx-test pg_catalog_version-test --gtest_filter PgCatalogVersionConnManagerTest.TestConnectionManagerChangePassword/1

./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestConnectionLimit
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestCorrectErrorMessage
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestDefaultConfig
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestDropAndRenameDb
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestDropAndRenameRole
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestEnquote
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestGFlags
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestLogicalClientVersion
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestMisc
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestPreparedStatements
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestSessionParameters
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestStatsAndMetrics
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestStickyConnections
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestUnixSocketConnections
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestUserContext
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestUserContextLimitedPoolSize
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestUserLoginLimit
./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestYCMConfiguration

Reviewers: kfranz, sanketh, skumar, mihnea

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D45699