[BACKPORT 2024.2][#27825] YSQL: Send AuthOK message received from Auth Backend immediately
Summary:
When we connect to an authentication backend, connection manager relays
messages back and forth from the server and external client, acting transparently.

This is not, however, the case when `AuthenticationOK` type message comes. This
is sent by Postgres at the end of the auth flow to indicate that the authentication
has been successfully performed. Instead of sending it back to the external client
directly, we are ignoring it and then sending it after `yb_auth_via_auth_backend`
returns.

There doesn't seem to be any reason for the current implementation. This diff
immediately relays back the AuthOK message to the client.

This is an innocuous change for now -- it was working earlier and will work now
(of course this seems easier to understand). However, when we decide to not buffer
the ParameterStatus packets from Postgres and instead relay them immediately to
the external client, this becomes a problem.

The Postgres driver expects AuthOK message first before receiving ParameterStatus
messages. When we buffer AuthOK and relay ParameterStatus, this confuses the driver
which errors out.
Jira: DB-17419

Original commit: d3c1775b6c86f48746f1eef97247d75b422532e9 / D45052

Test Plan:
Verified that the change works locally. Also ran a few Connection Manager Java tests
locally to verify that auth still works.

Jenkins: all tests

Reviewers: skumar, asrinivasan, rbarigidad, mkumar, vikram.damle

Reviewed By: vikram.damle

Subscribers: yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D45674