[BACKPORT 2025.1][#27339] YSQL: Fix create temp table with on commit drop
Summary:
Currently, creating a temporary table with the ON COMMIT DROP clause results in an error or crash when transactional DDL is not enabled.

If the command is issued within an explicit transaction, the `COMMIT` either fails with the error `Transaction for catalog table write operation 'pg_type' not found`, or crashes the backend if the table has any indexes.
If the command is not inside an explicit transaction, the `CREATE` itself fails with the same error or causes a crash.

In all cases, the root cause is that the DROP operation triggered by ON COMMIT DROP is not executed as part of a DDL transaction.
 - For standalone CREATE statements, the DDL transaction context has already ended by the time the drop is executed.
 - For explicit transactions, no DDL transaction is initiated during COMMIT, so the drop occurs outside of any DDL context.

To resolve the issue:
 - Call YBIncrementDdlNestingLevel() before dropping the relations marked for ON COMMIT DROP to start a DDL transaction.
 - Call YBDecrementDdlNestingLevel() after the drops to properly exit the autonomous DDL transaction.
 - Also, set the DDL mode to YB_DDL_MODE_SILENT_ALTERING, as only temporary tables (which are session-local) are affected.
Jira: DB-16848

Original commit: 9d4fd9b50a9a7796f1514ff084f414e41ed3fadc / D45645

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressFeature'

Reviewers: stiwary, kramanathan, #db-approvers

Reviewed By: kramanathan, #db-approvers

Subscribers: mihnea, svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D45739