[PLAT-18181] Added pagination support for getting user groups from Azure AD
Summary:
Currently, the code retrieves only the first page (100) of groups for a user from Azure AD. This is fixed by looping through all the pages.
Refer to: https://learn.microsoft.com/en-us/graph/paging?tabs=http

Test Plan:
Verified the current AD setup with a user with fewer than 100 groups to ensure the current behavior is intact
Created and assigned 262 groups to a user and ensured all groups are fetched for role assignment. One group was mapped to the admin role, and role bindings reflected correctly in the UI.

Reviewers: anabaria

Reviewed By: anabaria

Differential Revision: https://phorge.dev.yugabyte.com/D45725