[#28160] YSQL: Query Plan Management Infrastructure
Summary:
Query Plan Management (QPM) is a mechanism to monitor plan changes/usage over time. For example, when a user upgrades to a new release, or turns on the CBO, some plans may regress. In these cases, it is useful to have a way to allow users to detect and correct the plan regressions. For details on plan hints, please refer to [[ https://docs.yugabyte.com/stable/launch-and-manage/monitor-and-alert/query-tuning/pg-hint-plan/ | Optimize YSQL queries using pg_hint_plan]].

The feature is enabled/disabled by the GUC **//yb_pg_stat_plans_track//**, whose valid values are:

  - "none" : no QPM stats are collected (default)
  - "top" : only track top-level (non-nested) statements
  - "all" : track all statements

QPM uses a shared memory hash table that resides on each node (so all collected data is node-specific). The hash table contains the following fields:

  -  dbid (OID) : database id
  -   userid (OID) : user id
  -   queryid (BIGINT) : query id, a hash of the query representation
  -   planid (BIGINT) : plan id, a hash of the execution plan
  -   first_used (TIMESTAMPTZ : the time an entry is inserted
  -   last_used (TIMESTAMPTZ : the last time an entry is referenced
  -   hints (TEXT) : the plan hints that would generate the plan for the query
  -   calls (BIGINT) : the number of times an entry has been referenced
  -   avg_exec_time (DOUBLE PRECISION) : the average actual execution time for an entry
  -   max_exec_time (DOUBLE PRECISION) : the maximum execution time for the entry
  -   max_exec_time_params (TEXT) : the parameter values that gave the max (worst) execution time (if the query is parameterized)
  -   avg_est_cost (DOUBLE PRECISION) : the average estimated execution cost for the entry's plan
  -   plan (TEXT) : the EXPLAIN plan text for the entry (default format is JSON but this can be changed with a GUC)

The key for an entry is [dbid, userid, queryid, planid]. The statement types for which entries are created/updated are SELECT, INSERT, UPDATE, MERGE, DELETE, and EXECUTE. Entries are only created for these types if hints can be generated for them. (Yugabyte has a hint generation mechanism that, given a query plan, can generate the hints that will guarantee the same plan is generated for the statement. These generated hints are the ones stored in the QPM hash table.)

The default (fixed but configurable) size of the hash table is 5000 entries. If this limit is reached, the oldest entry is evicted. (By default, a simple clock LRU replacement algorithm is used for efficiency but a true LRU replacement strategy is available via a GUC.) The hash table entries are persisted across restarts.

Both hint and plan text have fixed sizes, 2048 and 4096 respectively. If the text does not fit, an attempt is made to compress it, If the compressed text is still too large, the original text is truncated. (If hints are truncated, they obviously cannot be used for plan pinning.)

Plan Id is a new value that is basically a hash computed from a query plan (similar to a Query Id, which is a hash value computed for a statement). The same statement can have multiple plans, e.g., if planner statistics change. EXPLAIN was also enhanced to both query id and plan id can be displayed for a plan:

```
EXPLAIN (queryid ON, planid ON, ...) <statement>;
```

Query text is not stored in the QPM hash table but can be found by joining the QPM table with the //**pg_stat_statements**// view on query id.

The primary way to access QPM data is by querying the //**yb_pg_stat_plans**// view:

```
SELECT * FROM  yb_pg_stat_plans;
```

There is also a view //**yb_pg_stat_plans_get_insights**// that provides information about plan/optimizer performance. This view identifies the plan with the lowest execution time for each query id. There are also fields that identify a plan if it has the lowest execution time but not the lowest estimated (planner) cost, or a plan that has the lowest estimated (planner) cost but not the lowest actual execution time.

There are new GUCs to control QPM behavior. Please refer to the QPM functional spec for descriptions.

Test Plan:
```
yb_build.sh --java-test 'org.yb.pgsql.TestYbQpm' : new Java test for QPM functionality
yb_build.sh --java-test 'org.yb.pgsql.TestYbPlanId' : new Java test for plan id
```

Reviewers: mihnea, kramanathan, jason, #db-approvers

Reviewed By: kramanathan, #db-approvers

Subscribers: svc_phabricator, smishra, jason, yql, kramanathan

Tags: #jenkins-ready, #jenkins-trigger

Differential Revision: https://phorge.dev.yugabyte.com/D45832