[BACKPORT 2025.1][#27942] CDC: Fix intent fetch for CDC with vector columns
Summary:
#####Backport Description
No merge conflicts were encountered.

#####Original Description
The revision D42176 introduced writing tombstone records to reverse vector index when vector is updated or removed. With this we started writing an additional intent corresponding to the transaction ID in which the vector column was updated or deleted. The key structure of this intent is like a reverse index intent (`kReverseTxnKey`). However unlike a reverse index intent, the value of this intent does not contain a key to main intent (i.e Prefix + SubDocKey (no HybridTime) + IntentType + HybridTime).

When reverse index iterator encounters such an intent while trying to fetch intents for a transaction for CDC, it tries to seek the intent iterator forward based on the value present in the tombstone intent event. But since the value of the tombstone intent is not a main intent key, this seek fails. As a result we do not send any intent belonging to this transaction to CDC producer and miss all the records present in such a transaction.

In this revision we fix this issue by skipping the intents corresponding to the vector tombstones. Skipping this intents will not lead to any loss of data, since the CDC receives data only from the main intents. These reverse index intents corresponding to vector tombstones do not point to any main intent, hence we can safely skip them for CDC.
Jira: DB-17557

Original commit: f1eb832b4b155e80e3f3efe5cc07dc79809f15ba / D45559

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testPgVectorWithLogicalReplication'

Reviewers: skumar, asrinivasan, sergei, vkushwaha

Reviewed By: vkushwaha

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45755