[#28129] DocDB, ASH: Fix metadata not passing with AcquireObjectLocks and ReleaseObjectLocks RPC
Summary:
After D44713, the threadlocal WaitStateInfoPtr needs to be
non empty and have the metadata for the new mechanism to
automatically pass ASH metadata through RPCs.

The AcquireObjectLocks and ReleaseObjectLocks RPCs are scheduled
from different threads, so we need to pass the WaitStateInfoPtr
across threads.

The WaitStateInfoPtr object is adopted when sending the RPC and not
earlier because this code is executed in master, and we don't set wait
events on master as of today. So it is fine not having a threadlocal
WaitStateInfoPtr for the entire duration. We only need to adopt this so
that the new mechanism to pass metadata can read from the
threadlocal WaitStateInfoPtr.

I also found that `ash::WaitStateInfo::SetCurrentWaitState(nullptr);`
was getting called from `WriteQuery::DoStartSynchronization` in this
code path, which makes the threadlocal WaitStateInfoPtr null even if we
adopt it early. So that is another reason to adopt the wait state pointer
this late.
Jira: DB-17769

Test Plan: ./yb_build.sh --cxx-test pg_ash-test --gtest_filter PgAshMasterMetadataSerializerTest.TestMasterMetadataSerializer

Reviewers: amitanand

Reviewed By: amitanand

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D45759