[PLAT-18232] When fetching gflags checksum, check the label type using the old/new helm naming style
Summary:
In a customer env which had a universe upgraded from 2.14, the code which fetches gflags checksum
was using new label `app.kubernetes.io/name` which was getting an NPE as this does not exist in the
universe.

Fix is to use the old label `app: <service-type>` for old naming style universes.

Test Plan: Added UT

Reviewers: anijhawan, anabaria

Reviewed By: anabaria

Differential Revision: https://phorge.dev.yugabyte.com/D45789