[BACKPORT 2.27.0.2337][PLAT-17904][PLAT-18078][dr][ddl] Exclude system_postgres from xCluster add/remove tables list
Summary:
This diff sets the database name based on the table id for the sequences_data table.

There is no postgres schema for that table and it belongs to the whole database and thus we put '_' as the pg schema.

Exclude the system_postgres database from the xCluster Add/Remove Tables modal by filtering it out from the list of tables displayed in the UI.

Additionally, update the logic to reflect the recent ybdb changes where ddl_queue and replicated_ddls are now treated as user tables. Ensure replicated_ddls remains hidden in the modal.

Original commits:
361c20bfeaf16479cd7cb9d9cb61b29eee37f3b4 / D45325 &
318fdebd8cbf782a2769c88b97003c1464f6eb9a / D45791

Test Plan:
Created an automatic mode DR config with multiple databases in replication and made sure the proper database name is shown for each sequences_data replication stream.

- Made sure the `system_postgres` database is not shown in the xCluster add/remove modal.
- Made sure the replicated_ddls table is hidden in the dr config page while ddl_queue table is properly shown.

Reviewers: amindrov, spothuraju, jmak, nbhatia, yng

Reviewed By: amindrov, spothuraju

Differential Revision: https://phorge.dev.yugabyte.com/D46203