[BACKPORT 2025.1][PLAT-17775][PLAT-18078][PLAT-17904][xCluster] DbScopedDdl: Update RPC calls to support xCluster with DDL replication setup
Summary:
This diff updates the yb-client code and yba code to add support for setting up and maintaining the config in YBA.

This diff sets the database name based on the table id for the sequences_data table. There is no postgres schema for that table and it belongs to the whole database and thus we put '_' as the pg schema.

Original commit: 23dbbb6afdefeafbe04e121450689dd63404e359 / D44539
361c20bfeaf16479cd7cb9d9cb61b29eee37f3b4 / D45325
318fdebd8cbf782a2769c88b97003c1464f6eb9a / D45791

Test Plan:
- Created a DR config with DDL replication enabled.
- Made sure the user is still able to fall back to the old DR mode using the runtime config `yb.xcluster.db_scoped.automatic_ddl.creationEnabled`
- Made sure the switchover, failover, and repair workflows are working.
- Created an automatic mode DR config with multiple databases in replication and made sure the proper database name is shown for each sequences_data replication stream.

Reviewers: amindrov, spothuraju, jmak, nbhatia, yng

Reviewed By: amindrov, jmak

Differential Revision: https://phorge.dev.yugabyte.com/D45890